backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
boto3==1.35.36
botocore==1.35.36
httpx==0.27.2
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Scheduling
apscheduler==3.10.4

# Testing
pytest==8.3.3
pytest-anyio==0.0.0
anyio[asyncio]==4.6.0
moto[ec2,s3,rds,elbv2,cloudwatch,sts,iam,ce]==5.0.14
